+++ src/main/java/Application/OpenCVServer/ServerRunner.java
@@ -1,13 +1,18 @@
 package Application.OpenCVServer;
 
+
 public class ServerRunner implements Runnable {
+    private OpenCVServerOperator activeOperator;
+
+    public ServerRunner() {
+        activeOperator = OpenCVServerOperator.getInstance();
+    }
     @Override
     public void run() {
         while(true) {
-            OpenCVServer activeServer = OpenCVServer.getInstance();
-            activeServer.startConnection();
-            activeServer.standbyForConnection();
-            activeServer.stopConnection();
+            activeOperator.startConnection();
+            activeOperator.standbyForConnection();
+            activeOperator.stopConnection();
         }
     }
 }
